move ChatMessage to DataObjects namespace
